chore: move CroppingProvider to components folder
